finish

+++ src/empire/object.c
@@ -311,7 +311,7 @@ void empire_object_init_cities(int empire_id)
             (city->type == EMPIRE_CITY_TRADE || city->type == EMPIRE_CITY_FUTURE_TRADE)) {
             obj->obj.trade_route_id = trade_route_new();
             array_item(objects, obj->obj.id + 1)->obj.trade_route_id = obj->obj.trade_route_id;
-            for (int j = obj->obj.id + 2; j < objects.size; j++) {
+            for (int j = obj->obj.id + 2; (unsigned int) j < objects.size; j++) {
                 full_empire_object *waypoint = array_item(objects, j);
                 if (waypoint->obj.type != EMPIRE_OBJECT_TRADE_WAYPOINT) {
                     break;
